+++ resources/perftest/perftest.cpp
@@ -93,17 +93,17 @@ class NewManagerInterface
 {
 public:
   NewManagerInterface(const std::unordered_map<std::string, std::string>& refToPrimStrs) : refToPrimStrs{refToPrimStrs} {}
-  std::shared_ptr<TraitsData> resolve(const std::string& ref, const TraitsData::TraitIds& traitIds);
+  std::shared_ptr<TraitsData> resolve(const std::string& ref, const TraitsData::TraitSet& traitSet);
 
 private:
   const std::unordered_map<std::string, std::string>& refToPrimStrs;
 };
 
-std::shared_ptr<TraitsData> NewManagerInterface::resolve(const std::string& ref, const TraitsData::TraitIds& traitIds)
+std::shared_ptr<TraitsData> NewManagerInterface::resolve(const std::string& ref, const TraitsData::TraitSet& traitSet)
 {
-  TraitsData::TraitIds populatedTraits;
+  TraitsData::TraitSet populatedTraits;
   bool getPath = false;
-  if(std::find(traitIds.cbegin(), traitIds.cend(), BlobTrait::kId) != traitIds.cend()) {
+  if(std::find(traitSet.cbegin(), traitSet.cend(), BlobTrait::kId) != traitSet.cend()) {
     populatedTraits.insert(BlobTrait::kId);
     getPath = true;
   }